Implement memset_explicit (C23).
This is just equivalent to memset in ORCA/C, since there are no optimizations that would remove memset operations. Note that in addition to clearing the old data in memory, this implementation of memset/memset_explicit also will not leave any of the old data in registers or in the stack space that it used.